kiwoom_mock_trader/requirements.txt
httpx>=0.27.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
pydantic>=2.7.0,<3.0.0
PyYAML>=6.0.1,<7.0.0
rich>=13.7.1,<14.0.0
pytest>=8.2.0,<9.0.0
websocket-client>=1.8.0,<2.0.0
tzdata>=2024.1
